fix date issue, pull availableActionDates into helper
today() was using current time which was causing it to be greater than the end date in cases where end date had the same date. Time does not matter in this case, so I am updating so today() returns todays date with time set to all 0's.

In further preparation of the start fresh command, I am getting rid of the availableActionDates and quantityPerDay getters.